+++ packages/package/test/fixtures/exports-replace/svelte.config.js
@@ -1,7 +1,0 @@
-const config = {
-	package: {
-		exports: () => false
-	}
-};
-
-export default config;