+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-svelte
 
+## 3.0.1
+
+### Patch Changes
+
+- fix: update dependency @sveltejs/package to v2.0.0 ([#9087](https://github.com/sveltejs/kit/pull/9087))
+
+## 3.0.0
+
+### Major Changes
+
+- breaking: update library scaffolding for `@sveltejs/package` version 2 ([#8922](https://github.com/sveltejs/kit/pull/8922))
+
 ## 2.3.4
 
 ### Patch Changes